Placing random tiles
Boy this would be prettier with the state monad...

I'm pretty sure this is the only reasonable way to represent it, since
the game state should include a random seed, but we can't place an
initial seed. Instead we generate the initial seed from the current
timestamp if not already present in the game state.