Randomly generate the initial grid
This took a good bit of restructuring. The first problem was that we
needed a random seed to create the initial game state (this was already
mostly true...)

The signal extras package provides the `foldp'` function, which allows a
function to be passed for the initial value, rather than a value itself,
and calls that function with the strictly fetched value of the signal.